remotewrite: Small refactor of directories.
In prep for adding v2 compliance support, some tiny clean up of dirs.
