+++ remotewrite/sender/cases/counter.go
