+++ remotewrite/sender/cases/gauge.go
