+++ remotewrite/sender/cases/histogram.go
